API IsAdmin middleware

+++ src/app/api/category/[id]/route.ts
@@ -1,7 +1,8 @@
 import Category from "@/models/Category";
 import connectDB from "@/config/connectDB";
 import slugify from "slugify";
-import { NextResponse } from "next/server";
+import { NextResponse, NextRequest } from "next/server";
+import { getUserFromRequest } from "@/lib/auth";
 
 export async function PATCH(
   req: Request,
@@ -10,6 +11,15 @@ export async function PATCH(
   await connectDB();
 
   try {
+    const { isAdmin } = await getUserFromRequest(req as NextRequest);
+
+    if (!isAdmin) {
+      return NextResponse.json(
+        { message: "Forbidden: Admins only" },
+        { status: 403 }
+      );
+    }
+
     const body = await req.json();
     const { name } = body;
     const { id } = await params;
@@ -59,6 +69,15 @@ export async function DELETE(
   await connectDB();
 
   try {
+    const { isAdmin } = await getUserFromRequest(req as NextRequest);
+
+    if (!isAdmin) {
+      return NextResponse.json(
+        { message: "Forbidden: Admins only" },
+        { status: 403 }
+      );
+    }
+
     const { id } = await params;
 
     const deletedCategory = await Category.findByIdAndDelete(id);
